Remove person concern and related db columns
