+++ app/models/user.rb
@@ -5,7 +5,10 @@ class User < ActiveRecord::Base
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
 
-  include PersonConcern
+   validates :email, presence: true, email: true
+   validates :first_name, presence: true
+   validates :last_name, presence: true
+
   has_many :addresses, inverse_of: :user
 
   # Patient associations
