+++ app/validators/email_validator.rb
@@ -1,0 +1,8 @@
+class EmailValidator < ActiveModel::EachValidator
+  def validate_each(record, attribute, value)
+    unless value =~ /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\z/i
+      record.errors[attribute] <<
+        (options[:message] || "is not a valid email address")
+    end
+  end
+end
