+++ db/migrate/20160810152112_drop_address_columns_from_users.rb
@@ -1,0 +1,10 @@
+class DropAddressColumnsFromUsers < ActiveRecord::Migration
+  def change
+    remove_column :users, :street
+    remove_column :users, :city
+    remove_column :users, :state
+    remove_column :users, :zip
+    remove_column :users, :latitude
+    remove_column :users, :longitude
+  end
+end
